Add base feeding logic

+++ src/EightCare.Domain/Common/Entity.cs
@@ -5,7 +5,7 @@ namespace EightCare.Domain.Common
     [ExcludeFromCodeCoverage]
     public abstract class Entity
     {
-        public long Id { get; protected set; }
+        public int Id { get; protected set; }
 
         public override bool Equals(object obj)
         {
